@@ -1,30 +1,26 @@
 {
-  "name": "parking-garage-client",
+  "name": "parking-garage-client-2",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@mui/icons-material": "^5.2.5",
-    "@mui/material": "^5.2.7",
-    "@mui/styled-engine": "^5.2.6",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^12.1.10",
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "@mui/icons-material": "^5.8.0",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.2.0",
+    "@testing-library/user-event": "^13.5.0",
     "axios": "^0.27.2",
-    "bcryptjs": "^2.4.3",
-    "chart.js": "^3.4.0",
-    "i18next": "^21.6.6",
-    "i18next-browser-languagedetector": "^6.1.2",
-    "react": "^17.0.2",
-    "react-chartjs-2": "^3.0.3",
-    "react-cookie": "^4.1.1",
-    "react-dom": "^17.0.2",
-    "react-i18next": "^11.15.3",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "4.0.3",
-    "sass": "^1.46.0",
-    "web-vitals": "^1.0.1"
+    "chart.js": "^3.8.0",
+    "i18next": "^21.8.4",
+    "i18next-browser-languagedetector": "^6.1.4",
+    "react": "^18.1.0",
+    "react-chartjs-2": "^4.1.0",
+    "react-dom": "^18.1.0",
+    "react-i18next": "^11.16.9",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
+    "sass": "^1.52.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
