@@ -1,77 +1,67 @@
 // REACT
-import { PropTypes } from 'prop-types'
-import React from 'react'
+import React, { useContext, useState } from 'react'
 // PLUGINS
 import axios from 'axios'
 // PACKAGES
 import i18n from '../plugins/i18n'
 // HELPERS
-import themeableClassName from '../helpers/themeableClassName'
+import { ThemeContext } from '../helpers/ThemeContext'
 // STYLES
 import '../stylesheets/Login.scss'
 
-class Login extends React.Component {
-  constructor(props) {
-    super(props)
+export const Login = (props) => {
+  const { closeWindow } = props
 
-    this.usernameInputRef = React.createRef()
-    this.passInputRef = React.createRef()
+  const [data, setData] = useState({username: '', password: ''})
+  const [error, setError] = useState()
 
-    this.state = {
-      error: null
-    }
+  const { colors } = useContext(ThemeContext)
+
+  const fieldChangeHandler = (e) => {
+    setData({ ...data, [e.target.name]: e.target.value })
   }
 
-  handleLoginForm(event) {
-    event.preventDefault()
-    const username = this.usernameInputRef.current.value
-    const password = this.passInputRef.current.value
+  const submitHandler = (e) => {
+    e.preventDefault()
 
-    axios.post(`http://aaronhost:8000/auth/authenticate?username=${username}&given_pswd=${password}`)
+    if (!data.username || !data.password) return
+
+    axios.post(`http://aaronhost:8000/auth/authenticate?username=${data.username}&given_pswd=${data.password}`)
       .then((response) => {
-        if (response.data && response.data.result) {
-          localStorage.setItem('jwt', response.data.result)
+        console.log('POST authenticate response', response)
+        if (response.statusText === 'OK') {
+          localStorage.setItem('jwt', response.data.json_web_token)
 
           window.location.pathname = '/profile'
         } else {
-          this.setState({ error: 'login_failed' })
+          setError('login_failed')
         }
       })
   }
 
-  handleEscPress(e) {
-    var { onClose } = this.props
-
-    if (e.key === 'Escape') {
-      onClose()
-    }
+  const escPressHandler = (e) => {
+    if (e.key === 'Escape') closeWindow()
   }
 
-  render() {
-    var { error } = this.state
-    var { currentTheme, onClose } = this.props
-
-    return (
-      <div
-        className={themeableClassName('Modal', currentTheme)}
-        onKeyDown={(e) => this.handleEscPress(e)}
-      >
-        <button onClick={onClose} className='Close'>&times;</button>
-        <h1>{i18n.t('Sign in')}</h1>
-        <form>
-          <span className='error-msg'>{i18n.t(error)}</span>
-          <input type='text' autoFocus={true} ref={this.usernameInputRef} placeholder={i18n.t('Username')} required />
-          <input type='password' ref={this.passInputRef} placeholder={i18n.t('Password')} required />
-          <button type='submit' onClick={(e) => this.handleLoginForm(e)}>{i18n.t('Sign in')}</button>
-        </form>
-      </div>
-    )
-  }
+  return (
+    <div
+      className='Modal'
+      onKeyDown={escPressHandler}
+      style={{ 
+        borderColor: colors.chartPrimary,
+        backgroundColor: colors.headerBackground
+       }}
+    >
+      <button onClick={closeWindow} className='Close'>
+        &times;
+      </button>
+      <h1 style={{ color: colors.font }}>{i18n.t('Sign in')}</h1>
+      <form onSubmit={submitHandler}>
+        <span className='error-msg'>{i18n.t(error)}</span>
+        <input type='text' name='username' onChange={fieldChangeHandler} autoFocus={true} placeholder={i18n.t('Username')} required />
+        <input type='password' name='password' onChange={fieldChangeHandler} placeholder={i18n.t('Password')} required />
+        <input type='submit' value={i18n.t('Sign in')} />
+      </form>
+    </div>
+  )
 }
-
-Login.propTypes = {
-  currentTheme: PropTypes.object.isRequired,
-  onClose: PropTypes.func.isRequired
-}
-
-export default Login
