@@ -1,133 +1,115 @@
 // REACT
-import { PropTypes } from 'prop-types'
-import React from 'react'
+import React, { useContext, useEffect, useState } from 'react'
 // PACKAGES
+import axios from 'axios'
 import i18n from '../plugins/i18n'
 // HELPERS
-import { authWithJwt } from '../helpers/auth'
-import themeableClassName from '../helpers/themeableClassName'
+import { ThemeContext } from '../helpers/ThemeContext'
 // SNIPPETS
 import states from '../snippets/usStates'
 // STYLES
 import '../stylesheets/EditProfile.scss'
 
-class EditProfile extends React.Component {
-  constructor(props) {
-    super(props)
+export const EditProfile = (props) => {
+  const { colors } = useContext(ThemeContext)
 
-    this.state = {
-      emailIsValid: false,
-      error: null,
-      inputs: [],
-      isLoaded: false,
-    }
+  const jsonWebToken = localStorage.getItem('jwt')
+  if (!jsonWebToken) { window.location.pathname = '/' }
+
+  const [emailIsValid, setEmailValid] = useState()
+
+  const [inputs, setInputs] = useState({})
+
+  const authUser = () => {
+    return axios.post('http://aaronhost:8000/auth/authorize', {}, {
+      headers: { 'Authorization': `Bearer ${jsonWebToken}` }
+    })
   }
 
-  async componentDidMount() {
-    let user = await authWithJwt()
-
-    if (!user.garage_id) {
-      window.location.pathname = '/'
-    }
-
-    fetch(`http://aaronhost:8000/garage/${user.garage_id}/profile`, {
-      method: 'GET'
-    })
-      .then(response => response.json())
-      .then(
-        (response) => {
-          this.setState({
-            emailIsValid: true,
-            inputs: response,
-            isLoaded: true
+  useEffect(() => {
+    authUser()
+      .then((response) => {
+        axios.get(`http://aaronhost:8000/garage/${response.data.garage_id}/profile`)
+          .then((response) => {
+            setEmailValid(true)
+            setInputs(response.data)
           })
-        },
-      )
+      })
+  }, [])
+
+  const inputChangeHandler = (e) => {
+    setInputs({ ...inputs, [e.target.name]: e.target.value })
   }
 
-  handleEmailValidation = (event) => {
-    let emailRegex = new RegExp(/^$|^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/)
-    var emailIsValid = emailRegex.test(event.target.value)
+  const emailValidationHandler = (e) => {
+    inputChangeHandler(e)
 
-    this.setState({ emailIsValid: emailIsValid })
+    const emailRegex = new RegExp(/^$|^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/)
+    const emailIsValid = emailRegex.test(inputs.email)
+
+    setEmailValid(emailIsValid)
   }
 
-  handleSubmit = (event) => {
-    event.preventDefault()
+  const submitHandler = (e) => {
+    e.preventDefault()
 
-    var jsonWebToken = localStorage.getItem('jwt')
+    console.log('submitHandler', new Date().toLocaleDateString('en-US', {
+      weekday: 'long',
+      year: 'numeric',
+      month: 'long',
+      day: 'numeric',
+      hour: 'numeric',
+      minute: 'numeric',
+      second: 'numeric'
+    }))
 
-    var inputs = {}
-    var inputParams
+    let inputParams = new URLSearchParams(inputs).toString()
 
-    for (var input of event.target) {
-      inputs[input['name']] = input['value']
-    }
-
-    inputParams = new URLSearchParams(inputs).toString()
-
-    fetch(`http://aaronhost:8000/garage/${inputs['garage_id']}/profile?${inputParams}`, {
-      method: 'PUT',
-      headers: new Headers({
-        'Content-Type': 'application/json',
-        'Authorization': 'Bearer ' + jsonWebToken
-      }),
-    })
-      .then(response => response.json())
-
-    window.location.pathname = '/'
+    authUser()
+      .then((response) => {
+        axios.put(`http://aaronhost:8000/garage/${response.data.garage_id}/profile?${inputParams}`, {
+          headers: new Headers({ 'Authorization': 'Bearer ' + response.data.jwt })
+        })
+          .then(window.location.pathname = `/${response.data.garage_id}`)
+          .catch(response => setInputs(response.data))
+      })
   }
 
-  render() {
-    var { emailIsValid, error, isLoaded, inputs } = this.state
-    var { currentTheme } = this.props
+  const labelStyle = {
+    color: colors.font
+  }
 
-    if (!error && isLoaded) {
-      return (
-        <form
-          className={themeableClassName('profile-form', currentTheme)}
-          onSubmit={this.handleSubmit}
-        >
-          <label>{i18n.t('Name of Garage')}</label>
-          <input name='garage_name' defaultValue={inputs['name']} />
-          <label>{i18n.t('Address 1')}</label>
-          <input name='address1' defaultValue={inputs['address1']} required />
-          <label>{i18n.t('Address 2')}</label>
-          <input name='address2' defaultValue={inputs['address2']} />
-          <label>{i18n.t('City')}</label>
-          <input name='city' defaultValue={inputs['city']} required />
-          <label>{i18n.t('State')}</label>
-          <select name='state' defaultValue={inputs['state']} required>
-            {states().map((state, i) =>
-              <option
-                key={i}
-                value={state.abbreviation}
-              >
-                {state.name}
-              </option>
-            )}
-          </select>
-          <label>{i18n.t('Zip Code')}</label>
-          <input name='zip' defaultValue={inputs['zip']} />
-          <label>{i18n.t('Email Address')}</label>
-          <input name='email' defaultValue={inputs['email']} onChange={this.handleEmailValidation} />
-          <input name='garage_id' defaultValue={inputs['id']} hidden={true} />
+  const inputStyle = {}
 
-          <input type='submit' disabled={!emailIsValid} />
-        </form>
-      )
-    } else {
-      return (
-        <div className='rendering-msg'>
-          ({error && error.message ? i18n.t('Service unavailable') : i18n.t('Loading...')})
-        </div>
-      )
-    }
+  if (inputs) {
+    return (
+      <form
+        className='profile-form'
+        onSubmit={submitHandler}
+      >
+        <label style={labelStyle}>{i18n.t('Name of Garage')}</label>
+        <input style={inputStyle} name='name' defaultValue={inputs.name} onChange={inputChangeHandler} />
+        <label style={labelStyle}>{i18n.t('Address 1')}</label>
+        <input style={inputStyle} name='address1' defaultValue={inputs.address1} onChange={inputChangeHandler} required />
+        <label style={labelStyle}>{i18n.t('Address 2')}</label>
+        <input style={inputStyle} name='address2' defaultValue={inputs.address2} onChange={inputChangeHandler} />
+        <label style={labelStyle}>{i18n.t('City')}</label>
+        <input style={inputStyle} name='city' defaultValue={inputs.city} onChange={inputChangeHandler} required />
+        <label style={labelStyle}>{i18n.t('State')}</label>
+        <select style={inputStyle} name='state' value={inputs.state} onChange={inputChangeHandler} required>
+          {states().map((state, i) =>
+            <option key={i} value={state.abbreviation}>
+              {state.name}
+            </option>
+          )}
+        </select>
+        <label style={labelStyle}>{i18n.t('Zip Code')}</label>
+        <input style={inputStyle} name='zip' defaultValue={inputs.zip} onChange={inputChangeHandler} />
+        <label style={labelStyle}>{i18n.t('Email Address')}</label>
+        <input style={inputStyle} name='email' defaultValue={inputs.email} onChange={emailValidationHandler} />
+
+        <input style={inputStyle} type='submit' disabled={!emailIsValid} />
+      </form>
+    )
   }
 }
-
-EditProfile.propTypes = {
-  currentTheme: PropTypes.object.isRequired
-}
-
-export default EditProfile
